relocate user access code
